[tests] Get rid of helper shell script for CMake build

+++ tests/lib/project.rb
@@ -132,7 +132,13 @@ class Project
 
     def cmake
         if @cmake.nil?
-            @cmake = 'cmake'
+            ENV['PATH'].split(':').each do |p|
+                exe = File.join(p, "cmake#{'.exe' if WINDOWS}")
+                next if !File.executable?(exe)
+                @cmake = exe
+                break
+            end
+            raise "Can't find CMake in PATH" if @cmake.nil?
         end
         @cmake
     end
@@ -173,8 +179,16 @@ class Project
         @properties['long'].to_s =~ /^(true|yes|1)/i ? true : false
     end
 
-    def has_build_script?
-        File.send(WINDOWS ? 'exists?' : 'executable?', File.join(path, 'build.sh'))
+    def has_script?(name, dir = nil)
+        File.send(WINDOWS ? 'exists?' : 'executable?', File.join(dir || path, name))
+    end
+
+    def has_onhost_script?(dir = nil)
+        has_script?('run-on-host', dir)
+    end
+
+    def has_build_script?(dir = nil)
+        has_script?('build', dir)
     end
 
     def has_cmakelists?(dir = nil)
@@ -378,35 +392,39 @@ class Project
             FileUtils.mkdir_p File.dirname(dir)
             FileUtils.cp_r path, dir
 
-            copy_cmakelists(dir) if has_cmakelists?
+            script = File.join(dir, 'run-on-host')
 
-            File.open(File.join(dir, 'run.sh'), 'w') do |bf|
-                bf.puts '#!/bin/sh'
-                bf.puts 'run()'
-                bf.puts '{'
-                bf.puts '    echo "## COMMAND: $@"'
-                bf.puts '    "$@"'
-                bf.puts '}'
-                if has_cmakelists?(dir)
-                    bf.puts "run rm -Rf #{dir}/cmake-build || exit 1"
-                    bf.puts "run mkdir -p #{dir}/cmake-build || exit 1"
-                    bf.puts "run cd #{dir}/cmake-build || exit 1"
-                    bf.puts "run cmake -DCMAKE_C_COMPILER=#{cc} -DCMAKE_CXX_COMPILER=#{cc} #{dir} || exit 1"
-                    bf.puts "run #{gnumake} -j#{@jobs} VERBOSE=1 || exit 1"
-                    bf.puts "run #{gnumake} test VERBOSE=1 || exit 1"
-                elsif File.exists?(File.join(dir, 'host', 'GNUmakefile'))
-                    bf.puts "run #{gnumake} -C #{File.join(dir, 'host')} -B -j#{@jobs} test CC=#{cc}"
-                else
-                    raise "Don't know how to run on-host testing for this test!"
+            if !has_onhost_script?
+                copy_cmakelists(dir)
+
+                File.open(script, 'w') do |bf|
+                    bf.puts '#!/bin/sh'
+                    bf.puts 'run()'
+                    bf.puts '{'
+                    bf.puts '    echo "## COMMAND: $@"'
+                    bf.puts '    "$@"'
+                    bf.puts '}'
+                    if has_cmakelists?(dir)
+                        bf.puts "run rm -Rf #{dir}/cmake-build || exit 1"
+                        bf.puts "run mkdir -p #{dir}/cmake-build || exit 1"
+                        bf.puts "run cd #{dir}/cmake-build || exit 1"
+                        bf.puts "run cmake -DCMAKE_C_COMPILER=#{cc} -DCMAKE_CXX_COMPILER=#{cc} #{dir} || exit 1"
+                        bf.puts "run #{gnumake} -j#{@jobs} VERBOSE=1 || exit 1"
+                        bf.puts "run #{gnumake} test VERBOSE=1 || exit 1"
+                    elsif File.exists?(File.join(dir, 'host', 'GNUmakefile'))
+                        bf.puts "exec #{gnumake} -C #{File.join(dir, 'host')} -B -j#{@jobs} test CC=#{cc}"
+                    else
+                        raise "Don't know how to run on-host testing for this test!"
+                    end
+                    bf.puts "exit 0"
                 end
-                bf.puts "exit 0"
+                FileUtils.chmod 0755, script
             end
-            FileUtils.chmod 0755, File.join(dir, 'run.sh')
 
             max_attempts = 5
             attempt = 1
             begin
-                run_cmd File.join(dir, 'run.sh'), errmsg: "On-host test of #{name} failed", track_mkdir_errors: true
+                run_cmd script, errmsg: "On-host test of #{name} failed", track_mkdir_errors: true
             rescue MkdirFailed
                 attempt += 1
                 raise "On-host testing of project #{name} failed" if attempt > max_attempts
@@ -443,96 +461,101 @@ class Project
         FileUtils.mkdir_p File.dirname(dstdir)
         FileUtils.cp_r path, dstdir
 
-        FileUtils.cd(dstdir) do
-            if has_cmakelists? && !has_build_script?
-                copy_cmakelists(dstdir)
+        bldscript = File.join(dstdir, 'build')
 
-                File.open(File.join(dstdir, 'build.sh'), 'w') do |bf|
-                    bf.puts '#!/bin/sh'
-                    bf.puts 'run()'
-                    bf.puts '{'
-                    bf.puts '    echo "## COMMAND: $@"'
-                    bf.puts '    "$@"'
-                    bf.puts '}'
+        if has_script?('build.sh') && !has_script?(File.basename(bldscript), dstdir)
+            FileUtils.cp File.join(dstdir, 'build.sh'), bldscript
+        end
 
-                    args = []
-                    args << "-DCMAKE_TOOLCHAIN_FILE=#{File.join(@ndk, 'cmake', 'toolchain.cmake')}"
-                    args << "-DANDROID_TOOLCHAIN_VERSION=#{@options[:toolchain_version]}" unless @options[:toolchain_version].nil?
-                    args << "-DANDROID_APP_PIE=#{options[:pie]}" unless options[:pie].nil?
+        copy_cmakelists(dstdir)
 
-                    %w[armeabi armeabi-v7a armeabi-v7a-hard x86 mips arm64-v8a x86_64 mips64].each do |abi|
-                        blddir = File.join(dstdir, 'cmake-build', abi)
-                        tmpinstalldir = File.join(dstdir, 'cmake-install', abi)
-                        installdir = File.join(dstdir, 'libs', abi)
+        errmsg = "Build of project #{name} failed"
 
-                        aargs = args.dup
-                        aargs << "-DCMAKE_INSTALL_PREFIX=#{tmpinstalldir}"
-                        aargs << "-DANDROID_ABI=#{abi}"
+        env = {}
+        env['V'] = '1'
+        env['APP_PIE'] = (options[:pie] ? true : false).to_s
 
-                        bf.puts "run rm -Rf #{blddir} #{tmpinstalldir} #{installdir} || exit 1"
-                        bf.puts "run mkdir -p #{blddir} || exit 1"
-                        bf.puts "run cd #{blddir} || exit 1"
-                        bf.puts "run #{cmake} #{aargs.join(' ')} #{dstdir} || exit 1"
-                        bf.puts "run #{gnumake} -j#{@jobs} VERBOSE=1 || exit 1"
-                        bf.puts "run #{gnumake} install VERBOSE=1 || exit 1"
-                        bf.puts "for f in `ls -1 #{tmpinstalldir}/bin/ 2>/dev/null` `ls -1 #{tmpinstalldir}/lib/lib*.so 2>/dev/null`; do"
-                        bf.puts "    mkdir -p #{installdir} || exit 1"
-                        bf.puts "    run cp $f #{installdir}/ || exit 1"
-                        bf.puts "done"
+        if has_script?(File.basename(bldscript), dstdir)
+            bldfunc = proc do
+                if WINDOWS
+                    shell = ENV['SHELL']
+                    if ENV['OSTYPE'] == 'cygwin'
+                        o,e,s = Open3.capture3("cygpath -m #{shell}")
+                        raise "Can't convert cygwin path to native: #{e}" unless s.success?
+                        shell = o.chomp
+                    else
+                        shell = shell.sub(/^\/([A-Za-z])\//, '\1:/')
                     end
-                    bf.puts "exit 0"
+                    cmd = "#{shell} #{bldscript}"
+                else
+                    cmd = bldscript
                 end
-                FileUtils.chmod 0755, File.join(dstdir, 'build.sh') unless WINDOWS
+                FileUtils.cd(dstdir) do
+                    run_cmd cmd, env: env, errmsg: errmsg, track_mkdir_errors: true
+                end
             end
+        elsif has_cmakelists?(dstdir)
+            bldfunc = proc do
+                args = [cmake]
+                args << "-DCMAKE_TOOLCHAIN_FILE=#{File.join(@ndk, 'cmake', 'toolchain.cmake')}"
+                args << "-DANDROID_TOOLCHAIN_VERSION=#{@options[:toolchain_version]}" unless @options[:toolchain_version].nil?
+                args << "-DANDROID_APP_PIE=#{options[:pie]}" unless options[:pie].nil?
+                %w[armeabi armeabi-v7a armeabi-v7a-hard x86 mips arm64-v8a x86_64 mips64].each do |abi|
+                    log_notice "BLD #{@display_type} [#{name}]#{variants(options)}: #{abi}"
 
-            bs = nil
-            [
-                File.join(dstdir, 'build.sh'),
-                @ndkbuild,
-            ].each do |bf|
-                next unless File.send(WINDOWS ? 'exists?' : 'executable?', bf)
-                bs = bf
-                break
+                    blddir = File.join(dstdir, 'cmake-build', abi)
+                    tmpinstalldir = File.join(dstdir, 'cmake-install', abi)
+                    installdir = File.join(dstdir, 'libs', abi)
+
+                    aargs = args.dup
+                    aargs << "-DCMAKE_INSTALL_PREFIX=#{tmpinstalldir}"
+                    aargs << "-DANDROID_ABI=#{abi}"
+                    aargs << dstdir
+
+                    FileUtils.rm_rf blddir
+                    FileUtils.rm_rf tmpinstalldir
+                    FileUtils.rm_rf installdir
+                    FileUtils.mkdir_p blddir
+
+                    FileUtils.cd(blddir) do
+                        run_cmd aargs.join(' '), errmsg: errmsg
+                        run_cmd "#{gnumake} -j#{@jobs} VERBOSE=1", errmsg: errmsg
+                        run_cmd "#{gnumake} install VERBOSE=1", errmsg: errmsg
+                    end
+
+                    bins = []
+                    bins += Dir.glob(File.join(tmpinstalldir, 'bin', '*')) if File.directory?(File.join(tmpinstalldir, 'bin'))
+                    bins += Dir.glob(File.join(tmpinstalldir, 'lib', 'lib*.so')) if File.directory?(File.join(tmpinstalldir, 'lib'))
+                    bins.each do |bin|
+                        FileUtils.mkdir_p installdir
+                        FileUtils.cp bin, installdir
+                    end
+                end
             end
-
-            raise "Don't know how to build project #{name}" if bs.nil?
-
-            env = {}
-            env['V'] = '1'
-            env['APP_PIE'] = (options[:pie] ? true : false).to_s
-
-            args = [bs]
-            if bs == @ndkbuild
-                args << "-B"
+        else
+            bldfunc = proc do
+                args = [@ndkbuild]
+                args << '-B'
                 args << "-j#{@jobs}"
                 env.each do |k,v|
-                    args << "#{k}=#{v}"
+                    args << "#{k}=#{v =~ /\s+/ ? "'#{v}'" : v}"
+                end
+                cmd = args.join(' ')
+                FileUtils.cd(dstdir) do
+                    run_cmd cmd, env: env, errmsg: errmsg, track_mkdir_errors: true
                 end
             end
+        end
 
-            cmd = args.join(' ')
-            if WINDOWS && bs != @ndkbuild
-                shell = ENV['SHELL']
-                if ENV['OSTYPE'] == 'cygwin'
-                    o,e,s = Open3.capture3("cygpath -m #{shell}")
-                    raise "Can't convert cygwin path to native: #{e}" unless s.success?
-                    shell = o.chomp
-                else
-                    shell = shell.sub(/^\/([A-Za-z])\//, '\1:/')
-                end
-                cmd = "#{shell} #{cmd}"
-            end
-
-            max_attempts = 5
-            attempt = 1
-            begin
-                run_cmd cmd, env: env, errmsg: "Build of project #{name} failed", track_mkdir_errors: true
-            rescue MkdirFailed
-                attempt += 1
-                raise "Build of project #{name} failed" if attempt > max_attempts
-                log_info "WARNING: Build of '#{name}' failed due to 'mkdir' error; trying again (attempt ##{attempt})"
-                retry
-            end
+        max_attempts = 5
+        attempt = 1
+        begin
+            bldfunc.call
+        rescue MkdirFailed
+            attempt += 1
+            raise "Build of project #{name} failed" if attempt > max_attempts
+            log_info "WARNING: Build of '#{name}' failed due to 'mkdir' error; trying again (attempt ##{attempt})"
+            retry
         end
 
         MRO.dump event: "build-success", path: path, pie: options[:pie]
